+++ test/models/reaction_test.exs
@@ -1,0 +1,18 @@
+defmodule Collaboration.ReactionTest do
+  use Collaboration.ModelCase
+
+  alias Collaboration.Reaction
+
+  @valid_attrs %{}
+  @invalid_attrs %{}
+
+  test "changeset with valid attributes" do
+    changeset = Reaction.changeset(%Reaction{}, @valid_attrs)
+    assert changeset.valid?
+  end
+
+  test "changeset with invalid attributes" do
+    changeset = Reaction.changeset(%Reaction{}, @invalid_attrs)
+    refute changeset.valid?
+  end
+end